% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_chars.R
\name{describe_chars}
\alias{describe_chars}
\title{Summarize the Non-Numeric Columns In A Dataframe}
\usage{
describe_chars(df)
}
\arguments{
\item{df}{a \code{dataframe}}
}
\description{
Summarize the Non-Numeric Columns In A Dataframe
}
