% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_numbers.R
\name{describe_numbers}
\alias{describe_numbers}
\title{Summarize the numeric values}
\usage{
describe_numbers(df)
}
\arguments{
\item{df}{A \code{dataframe}}
}
\description{
Summarize the numeric values
}
\examples{
#describe_numbers(df)
}
