\name{tellme-package}
\alias{tellme-package}
\alias{tellme}
\docType{package}
\title{\packageTitle{tellme}}
\description{\packageDescription{tellme}}
\section{Package Content}{\packageIndices{tellme}}
\author{\packageAuthor{tellme}}
\section{Maintainer}{\packageMaintainer{tellme}}
\keyword{package}
