% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_characters.R
\name{the_characters}
\alias{the_characters}
\title{Summarize the Characters}
\usage{
the_characters(df)
}
\arguments{
\item{df}{\code{dataframe} or \code{tibble}}
}
\value{
\code{tibble}, tibble of summary
}
\description{
This function provides a brief overview of the character columns in the supplied dataframe. With this function, you can easily discover the number of missing values and the number of unique values.
}
\examples{
# the_characters(iris)
}
