% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_numbers.R
\name{the_numbers}
\alias{the_numbers}
\title{Summarize the numeric values}
\usage{
the_numbers(df)
}
\arguments{
\item{df}{\code{dataframe} or \code{tibble}}
}
\value{
\code{tibble}, tibble of summary
}
\description{
This function is designed to provide a deeper insight into the numerical values. Think of this function as the base \code{summary} function on high-power. With this function, you'll be able to get the min, max, median, and different quartiles in one go.
}
\examples{
the_numbers(iris)
}
